backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
PyPDF2==3.0.1
scikit-learn==1.3.2
pandas==2.1.3
textblob==0.19.0
numpy==1.26.2
werkzeug==2.3.7
gunicorn==21.2.0
google-generativeai==0.3.2
python-dotenv==1.0.0 